:sparkles: feat: GamePulse SOD 서비스를 Clean Architecture로 마이그레이션 및 네임스페이스 업데이트
